refactor: Header 컴포넌트에서 status 분리
