+++ src/components/Header/Header.styles.tsx
@@ -42,14 +42,3 @@ export const ShoppingCartButtonText = styled.span`
   color: #fff;
   margin-right: 6px;
 `;
-
-export const ShoppingCartQuantity = styled.div`
-  background-color: #04c09e;
-  border-radius: 50%;
-  width: 26px;
-  height: 26px;
-  color: #fff;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
