+++ src/components/Header/Header.tsx
@@ -1,11 +1,8 @@
 import Logo from '../../assets/Logo.svg';
 import * as Styled from './Header.styles';
-import { useRecoilValue } from 'recoil';
-import { carListTotalQuantitySelector } from '../../stores/cartListStore.ts';
+import ShoppingCartStatus from './ShoppingCartStatus/ShoppingCartStatus.tsx';
 
 const Header = () => {
-  const totalNumber = useRecoilValue(carListTotalQuantitySelector);
-
   return (
     <Styled.HeaderContainer>
       <Styled.LogoButton>
@@ -14,9 +11,7 @@ const Header = () => {
       </Styled.LogoButton>
       <Styled.ShoppingCartButton>
         <Styled.ShoppingCartButtonText>장바구니</Styled.ShoppingCartButtonText>
-        <Styled.ShoppingCartQuantity>
-          <span>{totalNumber}</span>
-        </Styled.ShoppingCartQuantity>
+        <ShoppingCartStatus />
       </Styled.ShoppingCartButton>
     </Styled.HeaderContainer>
   );
