+++ src/components/Header/ShoppingCartStatus/ShoppingCartStatus.styles.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const ShoppingCartQuantity = styled.div`
+  background-color: #04c09e;
+  border-radius: 50%;
+  width: 26px;
+  height: 26px;
+  color: #fff;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
