+++ src/components/Header/ShoppingCartStatus/ShoppingCartStatus.tsx
@@ -1,0 +1,15 @@
+import * as Styled from './ShoppingCartStatus.styles.tsx';
+import { useRecoilValue } from 'recoil';
+import { carListTotalQuantitySelector } from '../../../stores/cartListStore.ts';
+
+const ShoppingCartStatus = () => {
+  const totalNumber = useRecoilValue(carListTotalQuantitySelector);
+
+  return (
+    <Styled.ShoppingCartQuantity>
+      <span>{totalNumber}</span>
+    </Styled.ShoppingCartQuantity>
+  );
+};
+
+export default ShoppingCartStatus;
